create the texture loader ext
