+++ examples/custom_sampler/application.rs
@@ -9,7 +9,7 @@
 
 use draw2d::{
     graphics::{
-        ext::SamplerFactory,
+        ext::{SamplerFactory, TextureLoader},
         layer::{Batch, LayerHandle},
         texture_atlas::TextureAtlas,
         vertex::Vertex2d,
@@ -53,10 +53,12 @@ impl Application {
     fn init(&mut self) -> Result<()> {
         self.update_projection();
 
-        let texture_handle_1 =
-            self.graphics.add_texture("assets/example.png")?;
-        let texture_handle_2 =
-            self.graphics.add_texture("assets/example.png")?;
+        let texture_handle_1 = self.graphics.add_texture(
+            self.graphics.read_texture_file("assets/example.png")?,
+        )?;
+        let texture_handle_2 = self.graphics.add_texture(
+            self.graphics.read_texture_file("assets/example.png")?,
+        )?;
 
         let sampler = unsafe {
             self.graphics.create_sampler(
