+++ src/graphics/ext/mod.rs
@@ -1,6 +1,8 @@
 mod sampler_factory;
 mod texture_2d_factory;
+mod texture_loader;
 
 pub use self::{
     sampler_factory::SamplerFactory, texture_2d_factory::Texture2dFactory,
+    texture_loader::TextureLoader,
 };
