+++ src/graphics/ext/sampler_factory.rs
@@ -1,7 +1,7 @@
-use crate::graphics::Graphics;
+use crate::graphics::{Device, Graphics};
 
-use anyhow::Result;
-use ash::vk;
+use anyhow::{Context, Result};
+use ash::{version::DeviceV1_0, vk};
 
 /// Types which implement this trait can create and destroy raw vulkan samplers.
 pub trait SamplerFactory {
@@ -38,5 +38,29 @@ impl SamplerFactory for Graphics {
 
     unsafe fn destroy_sampler(&self, sampler: vk::Sampler) {
         self.device.destroy_sampler(sampler)
+    }
+}
+
+impl SamplerFactory for Device {
+    unsafe fn create_sampler(
+        &self,
+        debug_name: impl Into<String>,
+        sampler_create_info: vk::SamplerCreateInfo,
+    ) -> Result<vk::Sampler> {
+        let owned_name = debug_name.into();
+        let sampler = self
+            .logical_device
+            .create_sampler(&sampler_create_info, None)
+            .with_context(|| {
+                format!("unable to create sampler {:?}", owned_name.clone())
+            })?;
+
+        self.name_vulkan_object(owned_name, vk::ObjectType::SAMPLER, &sampler)?;
+
+        Ok(sampler)
+    }
+
+    unsafe fn destroy_sampler(&self, sampler: vk::Sampler) {
+        self.logical_device.destroy_sampler(sampler, None);
     }
 }
