--- src/graphics/vulkan/device/ext/mod.rs
@@ -1,2 +1,0 @@
-mod sampler_factory_impl;
-mod texture_2d_factory_impl;
